requirements.txt
Pillow>=9.0.0
matplotlib>=3.5.0
google-genai>=0.2.0
openai>=1.0.0
python-dotenv>=0.21.0